+++ src/br/com/betha/controleacademico/view/ControleSuperior.java
@@ -18,6 +18,7 @@ import javax.swing.SwingConstants;
 import javax.swing.border.EmptyBorder;
 
 import br.com.betha.controleacademico.modelo.AlunoMedio;
+
 import java.awt.Window.Type;
 import java.awt.Dialog.ModalExclusionType;
 
@@ -99,8 +100,8 @@ public class ControleSuperior extends JFrame {
 								DateFormat formatter = new SimpleDateFormat("MM/dd/yyyy");   
 								
 								am.setPeriodo(txPeriodo.getText());
-								am.setPercentDesc(Integer.parseInt(txPercDesc.getText()));
-								am.setPercentAcresc(Integer.parseInt(txPercAcre.getText()));
+								am.setPercentDesc(Double.parseDouble(txPercDesc.getText()));
+								am.setPercentAcresc(Double.parseDouble(txPercAcre.getText()));
 								am.setObservacao(txObs.getText());
 								
 								//am.setVencMatricula(txVencimentoMat.getText());
